refactor: SurveyProgressBar 컴포넌트 생성
