+++ src/components/survey/content/SurveyContentItem.tsx
