+++ src/screens/survey/SurveyContentScreen.tsx
@@ -1,10 +1,11 @@
 import {useNavigation} from '@react-navigation/native';
-import {SurveyContentItem} from '@src/components/survey/SurveyContentItem';
-import {SurveyNextButton} from '@src/components/survey/SurveyNextButton';
+import {SurveyContentItem} from '@src/components/survey/content/SurveyContentItem';
+import {SurveyNextButton} from '@src/components/survey/common/SurveyNextButton';
 import {tw} from '@src/libs/tailwind';
 import {NavigationProps} from '@src/types/navigation';
 import React, {useState} from 'react';
 import {FlatList, Text, View} from 'react-native';
+import {SurveyProgressBar} from '@src/components/survey/common/SurveyProgressBar';
 
 export const SurveyContentScreen = () => {
   const navigation = useNavigation<NavigationProps>();
@@ -12,11 +13,7 @@ export const SurveyContentScreen = () => {
 
   return (
     <View style={tw`h-full w-full bg-white px-4 pt-10`}>
-      <View style={tw`flex flex-row items-center gap-1.5`}>
-        <View style={tw`h-2.5 flex-1 rounded-lg bg-primary-green`} />
-        <View style={tw`h-2.5 flex-1 rounded-lg bg-primary-green`} />
-        <View style={tw`h-2.5 flex-1 rounded-lg bg-custom-gray`} />
-      </View>
+      <SurveyProgressBar totalProgress={3} currentProgress={2} />
       <Text style={tw`pt-8 text-2xl font-bold text-custom-01`}>
         어떤 예능의 {/* TODO */}
       </Text>
